reset the wordCollection in redux when the user chooses a different verse
