+++ src/views/Study.js
@@ -16,6 +16,7 @@ import { useSelector, useDispatch } from "react-redux";
 import {setGlobalPhrase} from '../features/phraseSlice.js';
 import { Mic, MicOff, Check } from "lucide-react"
 import { setGlobalWordCollection, setWordCollectionInstance } from "../features/wordCollectionSlice.js";
+import createWordsCollection from '../utils/appUtils.js';
 
 // Mock scripture data (replace with actual data in a real application)
 const scripture = {
@@ -65,46 +66,6 @@ function a11yProps(index) {
     };
 }
 
-/**
- * This function takes a string of scripture text and splits it on a space.
- * It then loops through each word creating an object:
- *      If the "word" ends with punctuation, increment the phrase counter and add phrase to the object
- *      Add an index to the object
- *      Initialize "said" and "correct" both to boolean false
- *      add the word to a "word" property
- * 
- * @param {string} scriptureText
- * 
- * returns an array of word objects 
- */
-function createWordsCollection(scriptureText) {
-    let scriptureSimpleWordArray = scriptureText.split(/\s+/);
-    let scriptureWordsCollection = [];
-    let currentWord;
-    let currentWordObject;
-    let currentPhrase = 1;
-
-    for (let i=0; i < scriptureSimpleWordArray.length; i++) {
-        currentWord = scriptureSimpleWordArray[i];
-        currentWordObject = {};
-
-        currentWordObject.idx = i;
-        currentWordObject.word = currentWord;
-        currentWordObject.said = false;
-        currentWordObject.correct = null;
-
-        currentWordObject.phrase = currentPhrase;
-        
-        if (currentWord.match(/[.,/#!$%^&*;:{}=\-_`~()]/gu)) {
-            currentPhrase++
-        }
-
-        scriptureWordsCollection.push(currentWordObject)
-    }
-
-    return scriptureWordsCollection;
-}
-
 export default function Study() {
     const dispatch = useDispatch();
     const {
@@ -127,7 +88,7 @@ export default function Study() {
     scripture.replacedText = stripPunctuation(scripture.text)
     scripture.splitText = scripture.replacedText.split(/\s+/)
     const [activeTab, setActiveTab] = useState(0)
-    const [spokenText, setSpokenText] = useState("")
+    const [spokenText, setSpokenText] = useState(useSelector((state) => { return state.verse}))
     let [currentWordIndex, setCurrentWordIndex] = useState(0)
     let [wordCounter, setWordCounter] = useState(0)
     const [testSubmission, setTestSubmission] = useState("")
